Revert "Refactor configuration code"
