--- docker-compose.yml
@@ -1,36 +1,0 @@
-version: '3.9'
-services:
-  db:
-    image: postgres:16
-    environment:
-      POSTGRES_USER: mcp
-      POSTGRES_PASSWORD: mcp
-      POSTGRES_DB: mcp
-    ports:
-      - "5432:5432"
-  server:
-    build: .
-    depends_on:
-      - db
-    environment:
-      CONFIG_URL: /app/config.json
-      DB_URL: postgres://mcp:mcp@db:5432/mcp
-    volumes:
-      - ./fastmcp_server/config.json:/app/config.json:ro
-    ports:
-      - "3000:3000"
-    develop:
-      # Create a `watch` configuration to update the app
-      #
-      watch:
-        # Sync the working directory with the `/app` directory in the container
-        - action: sync
-          path: .
-          target: /app
-          # Exclude the project virtual environment
-          ignore:
-            - .venv/
-
-        # Rebuild the image on changes to the `pyproject.toml`
-        - action: rebuild
-          path: ./pyproject.toml
