--- fastmcp_server/.python-version
@@ -1,1 +1,0 @@
-3.13
