+++ fastmcp_server/config.json
@@ -4,11 +4,6 @@
       "path": "examples/swagger-pet-store.json",
       "apiBaseUrl": "https://petstore.swagger.io/v2",
       "prefix": "petstore"
-    },
-    {
-      "path": "https://raw.githubusercontent.com/xgirma/swagger.3.0.examples/refs/heads/master/link-example.json",
-      "apiBaseUrl": "https://petstore.swagger.io/v2/",
-      "prefix": "kingworks"
     }
   ],
   "server": {
