+++ fastmcp_server/requirements.txt
@@ -2,4 +2,3 @@ fastmcp
 uvicorn
 pytest
 pytest-httpx
-psycopg2-binary
