+++ fastmcp_server/tests/test_server.py
@@ -90,63 +90,3 @@ def test_list_server_endpoint():
     resp = asyncio.run(_call())
     assert resp.status_code == 200
     assert cfg["swagger"][0]["prefix"] in resp.json()["servers"]
-
-
-def test_load_multiple_configs(tmp_path):
-    cfg1 = {
-        "swagger": [{"path": "one.json", "apiBaseUrl": "http://a", "prefix": "a"}],
-        "server": {"host": "0.0.0.0", "port": 1},
-    }
-    cfg2 = {
-        "swagger": [{"path": "two.json", "apiBaseUrl": "http://b", "prefix": "b"}]
-    }
-    p1 = tmp_path / "c1.json"
-    p2 = tmp_path / "c2.json"
-    p1.write_text(json.dumps(cfg1))
-    p2.write_text(json.dumps(cfg2))
-    cfg = server.load_config([str(p1), str(p2)])
-    assert len(cfg["swagger"]) == 2
-
-
-def test_save_and_load_config_postgres(monkeypatch):
-    executed = {}
-
-    class FakeCursor:
-        def __init__(self):
-            self._result = [json.dumps({"swagger": [], "server": {"port": 5}})]
-
-        def execute(self, query, params=None):
-            executed.setdefault("queries", []).append((query, params))
-
-        def fetchone(self):
-            return [self._result[0]]
-
-        def __enter__(self):
-            return self
-
-        def __exit__(self, exc_type, exc, tb):
-            pass
-
-    class FakeConn:
-        def __enter__(self):
-            return self
-
-        def __exit__(self, exc_type, exc, tb):
-            pass
-
-        def cursor(self):
-            return FakeCursor()
-
-        def close(self):
-            executed["closed"] = True
-
-    def fake_connect(dsn):
-        executed["dsn"] = dsn
-        return FakeConn()
-
-    monkeypatch.setattr(server.db_utils.psycopg2, "connect", fake_connect)
-
-    cfg = {"swagger": [], "server": {"port": 5}}
-    server.save_config_to_postgres(cfg, "db")
-    loaded = server.load_config_from_postgres("db")
-    assert loaded["server"]["port"] == 5
