--- pyproject.toml
@@ -1,14 +1,0 @@
-[project]
-name = "mcp-manager"
-version = "0.1.0"
-description = "Add your description here"
-readme = "README.md"
-requires-python = ">=3.10"
-dependencies = [
-    "fastmcp>=2.8.1",
-    "ipykernel>=6.29.5",
-    "psycopg2-binary>=2.9.10",
-    "pytest>=8.4.1",
-    "pytest-httpx>=0.35.0",
-    "uvicorn>=0.34.3",
-]
